cc_requirements.txt
absl_py==2.1.0+computecanada
appdirs==1.4.4+computecanada
asttokens==2.4.1+computecanada
certifi==2024.2.2+computecanada
charset_normalizer==3.2.0+computecanada
chex==0.1.86+computecanada
click==8.1.7+computecanada
cloudpickle==3.0.0+computecanada
comm==0.2.2+computecanada
contextlib2==21.6.0+computecanada
contourpy==1.2.1+computecanada
cycler==0.12.1+computecanada
debugpy==1.8.1+computecanada
decorator==5.1.1+computecanada
distrax==0.1.2+computecanada
dm_control==1.0.15+computecanada
dm_env==1.6+computecanada
dm_tree==0.1.8+computecanada
docker-pycreds==0.4.0+computecanada
etils==1.7.0+computecanada
exceptiongroup==1.2.1+computecanada
executing==2.0.1+computecanada
Farama_Notifications==0.0.4+computecanada
flax==0.8.5+computecanada
fonttools==4.53.0+computecanada
fsspec==2024.9.0+computecanada
gast==0.6.0+computecanada
gitdb==4.0.11+computecanada
GitPython==3.1.43+computecanada
glfw @ file:///tmp/ebuser/avx512/GLFW/3.3.8/GCCcore-12.3-gentoo/glfw/glfw-2.6.2
gym==0.26.2+computecanada
gym_notices==0.0.8+computecanada
gymnasium==0.29.0+computecanada
idna==3.4+computecanada
imageio==2.34.1+computecanada
importlib_resources==6.0.1+computecanada
ipykernel==6.29.4+computecanada
ipython==8.25.0+computecanada
jax==0.4.30+computecanada
jaxlib==0.4.28+cuda12.cudnn89.computecanada
-e git+ssh://git@github.com/ikostrikov/jaxrl2.git@8b12c24543e14738974e9296f278aa6184e8fbb1#egg=jaxrl2
jedi==0.19.1+computecanada
jupyter_client==8.6.2+computecanada
jupyter_core==5.7.2+computecanada
kiwisolver==1.4.5+computecanada
labmaze==1.0.6+computecanada
lxml==5.2.2+computecanada
markdown_it_py==3.0.0+computecanada
matplotlib==3.9.0+computecanada
matplotlib_inline==0.1.7+computecanada
mdurl==0.1.2+computecanada
ml_collections==0.1.1+computecanada
ml_dtypes==0.4.0+computecanada
mpmath==1.3.0+computecanada
msgpack==1.0.8+computecanada
mujoco @ file:///tmp/ebuser/avx512/MuJoCo/3.1.6/GCCcore-12.3-gentoo/mujoco/mujoco-3.1.6
mujoco_mjx==3.1.6+computecanada
nest_asyncio==1.6.0+computecanada
nose==1.3.7+computecanada
numpy==1.26.4+computecanada
opencv_contrib_python==4.10.0
opencv_contrib_python_headless==4.10.0
opencv_python==4.10.0
opencv_python_headless==4.10.0
opt_einsum==3.3.0+computecanada
optax==0.2.2+computecanada
orbax_checkpoint==0.4.3+computecanada
packaging==24.1+computecanada
pandas==2.2.1+computecanada
parso==0.8.4+computecanada
pexpect==4.9.0+computecanada
pillow==10.3.0+computecanada
platformdirs==3.9.1+computecanada
prompt_toolkit==3.0.47+computecanada
protobuf==4.25.4+computecanada
psutil==5.9.8
ptyprocess==0.7.0+computecanada
pure_eval==0.2.2+computecanada
pygments==2.18.0+computecanada
PyOpenGL==3.1.7+computecanada
pyparsing==3.1.2+computecanada
python_dateutil==2.9.0.post0+computecanada
pytz==2024.1+computecanada
PyYAML==6.0.1+computecanada
pyzmq==26.0.3+computecanada
requests==2.31.0+computecanada
rich==13.8.1+computecanada
scipy==1.13.1+computecanada
seaborn==0.13.2+computecanada
sentry_sdk==2.2.1+computecanada
setproctitle==1.3.2+computecanada
six==1.16.0+computecanada
smmap==5.0.1+computecanada
stack_data==0.6.3+computecanada
sympy==1.12.1+computecanada
tensorboardX==2.6.2.2+computecanada
tensorflow_probability==0.21.0+computecanada
tensorstore==0.1.62+computecanada
termcolor==2.4.0+computecanada
toolz==0.12.1+computecanada
tornado==6.3.3+computecanada
tqdm==4.66.5+computecanada
traitlets==5.14.3+computecanada
typing_extensions==4.5.0+computecanada
tzdata==2024.1+computecanada
urllib3==2.1.0+computecanada
wandb==0.16.0+computecanada
wcwidth==0.2.13+computecanada
zipp==3.17.0+computecanada
